backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic-settings==2.0.3
cryptography==41.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
docker==6.1.3
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
